fix profile image when it's not uploaded

+++ src/main/java/com/project/marketplace/service/UserService.java
@@ -22,6 +22,7 @@ import java.io.File;
 import java.io.IOException;
 import java.util.Collections;
 import java.util.HashSet;
+import java.util.Objects;
 import java.util.Set;
 
 @Service
@@ -76,14 +77,14 @@ public class UserService implements UserDetailsService {
     @Transactional
     public void updateProfile(User user, MultipartFile multipartFile) throws IOException {
         User userFromDB = userRepository.findByNickname(user.getNickname());
-        if (multipartFile != null) {
+        if (multipartFile != null && !Objects.requireNonNull(multipartFile.getOriginalFilename()).isEmpty()) {
             if (userFromDB.getProfilePicture() != null) {
-                File lastPic = new File(userFromDB.getProfilePicture());
-                lastPic.delete();
+                File previousPic = new File(path + userFromDB.getProfilePicture());
+                previousPic.delete();
                 userRepository.deleteProfilePicture(userFromDB.getNickname());
             }
-            String fileName = path + user.getNickname() + "-profile-picture" + ".png";
-            multipartFile.transferTo(new File(fileName));
+            String fileName = user.getNickname() + "-profile-picture" + ".png";
+            multipartFile.transferTo(new File(path + fileName));
             user.setProfilePicture(fileName);
             userRepository.updateProfilePicture(fileName, user.getNickname());
         }
